+++ Earthfile
@@ -31,7 +31,7 @@ deps:
 code:
     FROM +deps
     COPY --platform=linux/amd64 ./ast/parser+parser/*.go ./ast/parser/
-    COPY --dir analytics autocomplete buildcontext builder cleanup cliutil cmd config conslogging debugger dockertar \
+    COPY --dir analytics autocomplete buildcontext builder cleanup cmd config conslogging debugger dockertar \
         docker2earthly domain logging secretsclient states util variables ./
     COPY --dir buildkitd/buildkitd.go buildkitd/settings.go buildkitd/
     COPY --dir earthfile2llb/*.go earthfile2llb/
