+++ analytics/analytics.go
@@ -16,7 +16,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/earthly/earthly/cliutil"
+	"github.com/earthly/earthly/util/cliutil"
 	"github.com/earthly/earthly/util/fileutil"
 	"github.com/earthly/earthly/util/gitutil"
 	"github.com/earthly/earthly/util/syncutil"
